+++ tests/test_central_full_session.py
@@ -37,7 +37,7 @@ class TestCentralFullSession:
         assert central.model == "PyDevCCU"
         assert central.get_client(interface_id=const.INTERFACE_ID).model == "PyDevCCU"
         assert central.primary_client.model == "PyDevCCU"
-        assert len(central._devices) == 395
+        assert len(central.devices) == 395
 
         data = {}
         for device in central.devices:
@@ -139,10 +139,6 @@ class TestCentralFullSession:
                 counter = usage_types[dp.usage]
                 usage_types[dp.usage] = counter + 1
 
-        addresses: dict[str, str] = {}
-        for address, device in central._devices.items():
-            addresses[address] = f"{device.model}.json"
-
         # check __dict__ / __slots__
         for device in central.devices:
             assert hasattr(device, "__dict__") is False
@@ -182,12 +178,12 @@ class TestCentralFullSession:
         assert len(data_point_types) == 6
         assert len(parameters) == 238
 
-        assert len(central._devices) == 395
+        assert len(central.devices) == 395
         virtual_remotes = ["VCU4264293", "VCU0000057", "VCU0000001"]
         await central.delete_devices(interface_id=const.INTERFACE_ID, addresses=virtual_remotes)
-        assert len(central._devices) == 392
+        assert len(central.devices) == 392
         del_addresses = list(central.device_descriptions.get_device_descriptions(interface_id=const.INTERFACE_ID))
         del_addresses = [adr for adr in del_addresses if ADDRESS_SEPARATOR not in adr]
         await central.delete_devices(interface_id=const.INTERFACE_ID, addresses=del_addresses)
-        assert len(central._devices) == 0
+        assert len(central.devices) == 0
         assert len(central.get_data_points(exclude_no_create=False)) == 0